+++ src/App.vue
@@ -1,44 +1,55 @@
 <template>
     <div id="app">
-        <Pagination
+        <vue-ads-pagination
             :total-items="200"
-            :max-visible-pages="4"
-            :page="3"
+            :max-visible-pages="5"
+            :page="page"
+            :loading="loading"
             @page-change="pageChange"
-            :detail-classes="['underline']"
-            :button-classes="buttonClasses"
-            :loading="true"
         >
             <template slot-scope="props">
-                Items {{ props.range.start }} tot {{ props.range.end }} van de {{ props.range.total }}
+                <div class="pr-2 leading-loose">
+                    Items {{ props.start }} tot {{ props.end }} van de {{ props.total }}
+                </div>
             </template>
-        </Pagination>
+            <template
+                slot="buttons"
+                slot-scope="props"
+            >
+                <vue-ads-page-button
+                    v-for="(button, key) in props.buttons"
+                    :key="key"
+                    v-bind="button"
+                    @page-change="page = button.page"
+                />
+            </template>
+        </vue-ads-pagination>
     </div>
 </template>
 
 <script>
-import Pagination from './components/Pagination';
+import '../node_modules/@fortawesome/fontawesome-free/css/all.css';
+import VueAdsPagination from './components/Pagination';
+import VueAdsPageButton from './components/PageButton';
 
 export default {
-    name: 'app',
+    name: 'App',
+
     components: {
-        Pagination,
+        VueAdsPageButton,
+        VueAdsPagination,
     },
 
     data () {
         return {
-            'buttonClasses': {
-                'default': ['border-none', 'bg-grey-lightest'],
-                'active': ['bg-orange', 'border-none'],
-                'dots': ['bg-white'],
-                'disabled': ['bg-grey-light'],
-            },
+            page: 5,
+            loading: false,
         };
     },
 
     methods: {
-        pageChange (page, range) {
-            console.log(page, range);
+        pageChange (page, start, end) {
+            console.log(page, start, end);
         },
     },
 };