+++ src/index.js
@@ -1,3 +1,8 @@
 import Pagination from './components/Pagination';
+import VueAdsPageButton from './components/PageButton';
 
 export default Pagination;
+
+export {
+    VueAdsPageButton,
+};