+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+#### v2.0.0 - 09/10/2018
+
+- Small rewrite of the paginator:
+    - Make it possible to use your own buttons or extend the default VueAdsPageButton, by adding a buttons template.
+    - So now we could remove the detailClasses and buttonClasses for custom styling.
+    - Remove the range object and add their attributes as clear properties. Only use start and end anymore.
+    - Purge the css.
+    - Move configs from package.json to separate files.
+
 #### v1.1.0 - 22/08/2018
 
 - Add a loading property to inidicate if the current page is loading